Update README.md
In dialogue with my senior thesis project, the final project for this class would be an interactive segment. I hope to create a poetic online webpage project, which use a series of photographs I made using iPhone during my independent studies period. Hopefully two question sections that give the audience a chance to enter "questions for Hunan" and "questions for the US".